Fix: Eliminar mensaje

+++ src/hooks/useAuthStore.js
@@ -1,6 +1,7 @@
 import { useDispatch, useSelector } from "react-redux";
 import { calendarApi } from "../api";
 import { clearErrorMessage, onChecking, onLogin, onLogout, onLogoutCalendar } from "../store";
+import Swal from "sweetalert2/dist/sweetalert2.all";
 
 export const useAuthStore = () => {
 
@@ -54,7 +55,7 @@ export const useAuthStore = () => {
         const tokenStorage = localStorage.getItem("token");
 
         try {
-            if (!tokenStorage) return dispatch(onLogout("La sesión expiró vuelva a autenticarse"));
+            if (!tokenStorage) return dispatch(onLogout());
             const data = await calendarApi.get("/auth/reset");
             const { fullname, token } = data.data;
 
@@ -69,9 +70,29 @@ export const useAuthStore = () => {
     }
 
     const startLogout = () => {
-        localStorage.clear();
-        dispatch(onLogoutCalendar())
-        dispatch(onLogout())
+        Swal.fire({
+            title: "Cerrar sesión?",
+            text: "¡Revisa guardar todo!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Confirmar",
+            cancelButtonText: "Cancelar"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                localStorage.clear();
+                dispatch(onLogoutCalendar())
+                dispatch(onLogout())
+                Swal.fire({
+                    title: "Sesión cerrada",
+                    text: "¡Hasta pronto!",
+                    icon: "success"
+                });
+            }
+        });
+
+
     }
     return {
         // * propiedades
